+++ src/bioimageio/spec/summary.py
@@ -1,4 +1,6 @@
+import html
 import os
+import platform
 import subprocess
 from dataclasses import dataclass
 from datetime import datetime, timezone
@@ -56,6 +58,8 @@ from ._internal.warning_levels import (
 )
 from ._version import VERSION
 from .conda_env import CondaEnv
+
+CONDA_CMD = "conda.bat" if platform.system() == "Windows" else "conda"
 
 Loc = Tuple[Union[int, str], ...]
 """location of error/warning in a nested data structure"""
@@ -210,10 +214,10 @@ class ValidationDetail(BaseModel, extra="allow"):
                         write_yaml(dumped_env, f)
 
                     compare_proc = subprocess.run(
-                        ["conda", "compare", str(path)],
+                        [CONDA_CMD, "compare", str(path)],
                         stdout=subprocess.PIPE,
                         stderr=subprocess.STDOUT,
-                        shell=True,
+                        shell=False,
                         text=True,
                     )
                     self.saved_conda_compare = (
@@ -280,10 +284,10 @@ class ValidationSummary(BaseModel, extra="allow"):
     def conda_list(self):
         if self.saved_conda_list is None:
             p = subprocess.run(
-                ["conda", "list"],
+                [CONDA_CMD, "list"],
                 stdout=subprocess.PIPE,
                 stderr=subprocess.STDOUT,
-                shell=True,
+                shell=False,
                 text=True,
             )
             self.saved_conda_list = (
@@ -635,14 +639,14 @@ def _format_summary(
             ["status", summary.status],
         ]
         if not hide_source:
-            info_rows.append(["source", summary.source_name])
+            info_rows.append(["source", html.escape(summary.source_name)])
 
         if summary.id is not None:
             info_rows.append(["id", summary.id])
 
         info_rows.append(["format version", f"{summary.type} {summary.format_version}"])
         if not hide_env:
-            info_rows.extend([[e.name, e.version] for e in summary.env])
+            info_rows.extend([[e.name, e.version] for e in sorted(summary.env)])
 
         if include_conda_list:
             info_rows.append(